refactor(codegen): split bindings/text-blocks/overflow into codegen/ submodules
Phase 1 Batch 3 (Moves 7-9 from docs/PHASE1-PLAN.md).  Extract 15
handlers plus the four overflow-warning helpers from latex_gen.py
into three new mixin modules.  The smoke-test mixin's PageBreak
handler is absorbed into text_blocks.py at the same time.

Moves are byte-identical Fowler Move-Method / Extract-Class
refactorings.  Method bodies, signatures, names, and docstrings are
unchanged.

* bindings.py — Theta, Binding (the ``{|...|}`` literal).  The
  multi-typed comprehension that uses bindings ships with the
  expressions mixin as part of _generate_set_comprehension.
* text_blocks.py — Section, Solution, Part, Paragraph (TEXT:),
  PureParagraph (PURETEXT:), LatexBlock (LATEX:), RawLatexBlock,
  BMachine (B:), LineBreak, PageBreak (relocated from _smoke.py),
  Contents, PartsFormat, TruthTable.
* overflow.py — _check_overflow, emit_warnings, get_warnings,
  clear_warnings, DEFAULT_OVERFLOW_THRESHOLD.  State attributes
  (_warn_overflow, _overflow_warnings, _overflow_threshold) remain
  initialised in LaTeXGenerator.__init__; only the methods move.
* _smoke.py — reduced to a placeholder class.  The PageBreak
  handler that smoke-tested the mixin pattern in Move 0 is now in
  text_blocks.py where it belongs.

The CodegenDispatch TYPE_CHECKING block grows again to declare the
new cross-cutting helpers (_escape_latex, _escape_latex_text,
_process_paragraph_text, _convert_operators_to_latex,
_generate_document_items_with_consolidation, _emit_attr_name) and
the per-instance state attributes (_warn_overflow,
_overflow_warnings, _overflow_threshold, _first_part_in_solution,
parts_format, toc_parts).

pyproject.toml — codegen/overflow.py joins the T20 allowlist for
its emit_warnings stderr print, matching the existing pattern for
cli.py, compile.py, and repl.py.

Gate 3 verified:
* make check — lint + markdown + format + mypy + pyright + 4700
  unit tests, all clean.
* make test-e2e — 159/159 examples/ .tex fixtures byte-identical.
* make refactor-diff — 190/190 examples/ + tests/bugs/ outputs
  byte-identical against main.

latex_gen.py shrinks from 3400 → 2871 lines.  Cumulative reduction
from the original 6864-line monolith: 58.2%.
