+++ src/txt2tex/codegen/__init__.py
@@ -5,9 +5,12 @@ from __future__ import annotations
 from txt2tex.codegen import (
     _smoke,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     algebra,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    bindings,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     expressions,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    overflow,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     paragraphs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     proofs,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     schemas,  # noqa: F401  # pyright: ignore[reportUnusedImport]
+    text_blocks,  # noqa: F401  # pyright: ignore[reportUnusedImport]
     types,  # noqa: F401  # pyright: ignore[reportUnusedImport]
 )
